Fixed the usage of the Win32 threads.
The misuse of API could cause a crash.

+++ threads/rt_thread_pool.c
@@ -76,11 +76,11 @@ rt_thread_pool_t *rt_tp_init(size_t number_of_threads)
 
 void rt_tp_deinit(rt_thread_pool_t *thread_pool)
 {
-    for (int i = 0; i < thread_pool->number_of_threads; ++i)
+    for (size_t i = 0; i < thread_pool->number_of_threads; ++i)
     {
         enqueue_work(thread_pool, RT_TP_WORK_TYPE_STOP, NULL, NULL, NULL);
     }
-    for (int i = 0; i < thread_pool->number_of_threads; ++i)
+    for (size_t i = 0; i < thread_pool->number_of_threads; ++i)
     {
         rt_thread_join(thread_pool->threads[i]);
     }
